Split main tests to separate file
